refactor 관리자페이지 상품 리스트 섹션 데이터/UI 로직 분리 && 관련 기능 별로 나눔 (+주석처리)
